backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic-settings==2.0.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
